requirements.txt
# Core Discord Bot
discord.py>=2.3.2
nextcord>=2.6.0

# LLM Providers
google-generativeai>=0.3.2
openai>=1.12.0
anthropic>=0.18.0
ollama>=0.1.6

# HTTP & Async
aiohttp>=3.9.3
httpx>=0.26.0

# Configuration & Database
pyyaml>=6.0.1
python-dotenv>=1.0.1
tinydb>=4.8.0

# Web Search
google-api-python-client>=2.116.0
duckduckgo-search>=4.1.1
requests>=2.31.0

# GUI (Optional - choose one or both)
customtkinter>=5.2.1
flask>=3.0.2
flask-cors>=4.0.0
flask-socketio>=5.3.6

# MCP Integration
mcp>=0.1.0

# Utilities
pydantic>=2.6.1
python-dateutil>=2.8.2
colorama>=0.4.6
rich>=13.7.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.4
black>=24.1.1
